category/games/2026/02/20260209-143522-vector-super-mario-bros-jump-infinite/pyproject.toml
[project]
name = "vector-super-mario-bros-jump-infinite"
version = "0.1.0"
description = "A precision-based vertical platformer inspired by Super Mario's classic jump physics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
